refactor(api): move machine Redfish helpers into machine module
Relocate machine-state-controller-specific Redfish utilities from the shared API
Redfish module into `state_controller::machine::redfish`.

This keeps host power control and DPU boot progress handling closer to the
machine state handler, reducing API crate coupling as preparation for extracting
the machine state controller into its own crate.
